+++ api/api-data-sharing-usage-data-v1/src/test/groovy/com/thoughtworks/go/apiv1/datasharing/usagedata/UsageStatisticsControllerV1DelegateTest.groovy
@@ -14,11 +14,10 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.apiv1.datasharing
+package com.thoughtworks.go.apiv1.datasharing.usagedata
 
 import com.thoughtworks.go.api.SecurityTestTrait
 import com.thoughtworks.go.api.spring.ApiAuthenticationHelper
-import com.thoughtworks.go.apiv1.datasharing.usagedata.UsageStatisticsControllerV1Delegate
 import com.thoughtworks.go.apiv1.datasharing.usagedata.representers.UsageStatisticsRepresenter
 import com.thoughtworks.go.server.domain.UsageStatistics
 import com.thoughtworks.go.server.service.DataSharingService
@@ -26,11 +25,13 @@ import com.thoughtworks.go.spark.AdminUserSecurity
 import com.thoughtworks.go.spark.ControllerTrait
 import com.thoughtworks.go.spark.NormalUserSecurity
 import com.thoughtworks.go.spark.SecurityServiceTrait
+import com.thoughtworks.go.server.util.RSAEncryptionHelper
 import com.thoughtworks.go.util.SystemEnvironment
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Nested
 import org.junit.jupiter.api.Test
 import org.mockito.Mock
+import org.skyscreamer.jsonassert.JSONAssert
 
 import static com.thoughtworks.go.api.base.JsonUtils.toObjectString
 import static org.mockito.Mockito.when
@@ -85,7 +86,7 @@ class UsageStatisticsControllerV1DelegateTest implements SecurityServiceTrait, C
         assertThatResponse()
           .isOk()
           .hasContentType(controller.mimeType)
-          .hasBody(toObjectString({ UsageStatisticsRepresenter.toJSON(it, metrics, true) }))
+          .hasBody(toObjectString({ UsageStatisticsRepresenter.toJSON(it, metrics) }))
       }
     }
   }
@@ -123,14 +124,15 @@ class UsageStatisticsControllerV1DelegateTest implements SecurityServiceTrait, C
         when(systemEnvironment.getUpdateServerPublicKeyPath()).thenReturn(publicKeyFile.getAbsolutePath())
         when(dataSharingService.getUsageStatistics()).thenReturn(metrics)
 
-        def expectedJson = toObjectString({ UsageStatisticsRepresenter.toJSON(it, metrics, false) })
+        def expectedJson = toObjectString({ UsageStatisticsRepresenter.toJSON(it, metrics) })
 
         getWithApiHeader(controller.controllerPath('/encrypted'))
 
+        def actualEncrypted = response.getContentAsString()
+        JSONAssert.assertEquals(RSAEncryptionHelper.decrypt(actualEncrypted, privateKeyFile.getAbsolutePath()), expectedJson, true)
         assertThatResponse()
           .isOk()
           .hasContentType("application/octet-stream")
-          .hasEncryptedBody(expectedJson, privateKeyFile.getAbsoluteFile())
       }
     }
   }
