+++ base/build.gradle
@@ -28,8 +28,6 @@ dependencies {
   compile group: 'commons-logging', name: 'commons-logging', version: '1.2'
   compile group: 'org.slf4j', name: 'slf4j-api', version: project.versions.slf4j
   compile group: 'ch.qos.logback', name: 'logback-classic', version: project.versions.logback
-  compile group: 'org.bouncycastle', name: 'bcprov-jdk15on', version: '1.47'
-
   testCompileOnly group: 'junit', name: 'junit', version: project.versions.junit
   testRuntimeOnly group: 'org.junit.vintage', name: 'junit-vintage-engine', version: project.versions.junit5
   testImplementation group: 'org.junit.jupiter', name: 'junit-jupiter-api', version: project.versions.junit5
