+++ server/src/main/java/com/thoughtworks/go/server/util/RSAEncryptionHelper.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.thoughtworks.go.util;
+package com.thoughtworks.go.server.util;
 
 
 import org.bouncycastle.util.io.pem.PemReader;
