+++ test/http-mocks/src/main/java/com/thoughtworks/go/http/mocks/MockHttpServletResponseAssert.java
@@ -16,7 +16,6 @@
 
 package com.thoughtworks.go.http.mocks;
 
-import com.thoughtworks.go.util.RSAEncryptionHelper;
 import net.javacrumbs.jsonunit.fluent.JsonFluentAssert;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.ReflectionToStringBuilder;
@@ -28,7 +27,6 @@ import org.assertj.core.util.Objects;
 import javax.activation.MimeType;
 import javax.activation.MimeTypeParseException;
 import javax.servlet.http.Cookie;
-import java.io.File;
 import java.io.UnsupportedEncodingException;
 
 import static javax.servlet.http.HttpServletResponse.SC_REQUEST_ENTITY_TOO_LARGE;
@@ -115,12 +113,6 @@ public class MockHttpServletResponseAssert<SELF extends MockHttpServletResponseA
             this.as("body");
             throw Failures.instance().failure(info, shouldBeEqual(actual.getContentAsByteArray(), expected, info.representation()));
         }
-        return myself;
-    }
-
-    public SELF hasEncryptedBody(String expected, File privateKeyFile) throws Exception {
-        String decrypted = RSAEncryptionHelper.decrypt(actual.getContentAsString(), privateKeyFile.getAbsolutePath());
-        JsonFluentAssert.assertThatJson(decrypted).isEqualTo(expected);
         return myself;
     }
